refactor(ui): extract view-model/coordinator-approval.ts
Move deriveCoordinatorApprovalSummary, shouldShowCoordinatorReviewAction,
and summarizeSyncRunResult into view-model/coordinator-approval.ts. The
barrel re-exports all three so the call sites in team-sync and the
sync event handlers stay on the same import path.
